requirements.txt
requests>=2.31.0
yfinance>=0.2.36
apscheduler>=3.10.4
streamlit>=1.32.0
python-dotenv>=1.0.1
